engine: spread functionality to several files
