+++ engine/engine.go
@@ -18,14 +18,12 @@ import (
 	"github.com/go-kit/log/level"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promauto"
-	"github.com/prometheus/prometheus/model/labels"
 	"github.com/prometheus/prometheus/promql"
 	"github.com/prometheus/prometheus/promql/parser"
 	"github.com/prometheus/prometheus/storage"
 	"github.com/prometheus/prometheus/util/stats"
 	v1 "github.com/prometheus/prometheus/web/api/v1"
 
-	"github.com/thanos-io/promql-engine/api"
 	"github.com/thanos-io/promql-engine/execution"
 	"github.com/thanos-io/promql-engine/execution/function"
 	"github.com/thanos-io/promql-engine/execution/model"
@@ -61,11 +59,6 @@ type Opts struct {
 	// in the new engine, instead of falling back to prometheus engine.
 	DisableFallback bool
 
-	// DebugWriter specifies output for debug (multi-line) information meant for humans debugging the engine.
-	// If nil, nothing will be printed.
-	// NOTE: Users will not check the errors, debug writing is best effort.
-	DebugWriter io.Writer
-
 	// ExtLookbackDelta specifies what time range to use to determine valid previous sample for extended range functions.
 	// Defaults to 1 hour if not specified.
 	ExtLookbackDelta time.Duration
@@ -98,78 +91,6 @@ func (o Opts) getLogicalOptimizers() []logicalplan.Optimizer {
 		copy(optimizers, o.LogicalOptimizers)
 	}
 	return optimizers
-}
-
-type remoteEngine struct {
-	q         storage.Queryable
-	engine    *compatibilityEngine
-	labelSets []labels.Labels
-	maxt      int64
-	mint      int64
-}
-
-func NewRemoteEngine(opts Opts, q storage.Queryable, mint, maxt int64, labelSets []labels.Labels) *remoteEngine {
-	return &remoteEngine{
-		q:         q,
-		labelSets: labelSets,
-		maxt:      maxt,
-		mint:      mint,
-		engine:    New(opts),
-	}
-}
-
-func (l remoteEngine) MaxT() int64 {
-	return l.maxt
-}
-
-func (l remoteEngine) MinT() int64 {
-	return l.mint
-}
-
-func (l remoteEngine) LabelSets() []labels.Labels {
-	return l.labelSets
-}
-
-func (l remoteEngine) NewRangeQuery(ctx context.Context, opts promql.QueryOpts, qs string, start, end time.Time, interval time.Duration) (promql.Query, error) {
-	return l.engine.NewRangeQuery(ctx, l.q, opts, qs, start, end, interval)
-}
-
-type distributedEngine struct {
-	endpoints    api.RemoteEndpoints
-	remoteEngine *compatibilityEngine
-}
-
-func NewDistributedEngine(opts Opts, endpoints api.RemoteEndpoints) v1.QueryEngine {
-	opts.LogicalOptimizers = []logicalplan.Optimizer{
-		logicalplan.PassthroughOptimizer{Endpoints: endpoints},
-		logicalplan.DistributeAvgOptimizer{},
-		logicalplan.DistributedExecutionOptimizer{Endpoints: endpoints},
-	}
-
-	return &distributedEngine{
-		endpoints:    endpoints,
-		remoteEngine: New(opts),
-	}
-}
-
-func (l distributedEngine) SetQueryLogger(log promql.QueryLogger) {}
-
-func (l distributedEngine) NewInstantQuery(ctx context.Context, q storage.Queryable, opts promql.QueryOpts, qs string, ts time.Time) (promql.Query, error) {
-	// Truncate milliseconds to avoid mismatch in timestamps between remote and local engines.
-	// Some clients might only support second precision when executing queries.
-	ts = ts.Truncate(time.Second)
-
-	return l.remoteEngine.NewInstantQuery(ctx, q, opts, qs, ts)
-}
-
-func (l distributedEngine) NewRangeQuery(ctx context.Context, q storage.Queryable, opts promql.QueryOpts, qs string, start, end time.Time, interval time.Duration) (promql.Query, error) {
-	// Truncate milliseconds to avoid mismatch in timestamps between remote and local engines.
-	// Some clients might only support second precision when executing queries.
-	start = start.Truncate(time.Second)
-	end = end.Truncate(time.Second)
-	interval = interval.Truncate(time.Second)
-
-	return l.remoteEngine.NewRangeQuery(ctx, q, opts, qs, start, end, interval)
 }
 
 func New(opts Opts) *compatibilityEngine {
@@ -231,7 +152,6 @@ func New(opts Opts) *compatibilityEngine {
 		prom:      engine,
 		functions: functions,
 
-		debugWriter:       opts.DebugWriter,
 		disableFallback:   opts.DisableFallback,
 		logger:            opts.Logger,
 		lookbackDelta:     opts.LookbackDelta,
@@ -250,8 +170,6 @@ func New(opts Opts) *compatibilityEngine {
 type compatibilityEngine struct {
 	prom      v1.QueryEngine
 	functions map[string]*parser.Function
-
-	debugWriter io.Writer
 
 	disableFallback   bool
 	logger            log.Logger
@@ -312,18 +230,13 @@ func (e *compatibilityEngine) NewInstantQuery(ctx context.Context, q storage.Que
 		return nil, err
 	}
 
-	if e.debugWriter != nil {
-		explain(e.debugWriter, exec, "", "")
-	}
-
 	return &compatibilityQuery{
-		Query:       &Query{exec: exec, opts: opts},
-		engine:      e,
-		expr:        expr,
-		ts:          ts,
-		t:           InstantQuery,
-		resultSort:  resultSort,
-		debugWriter: e.debugWriter,
+		Query:      &Query{exec: exec, opts: opts},
+		engine:     e,
+		expr:       expr,
+		ts:         ts,
+		t:          InstantQuery,
+		resultSort: resultSort,
 	}, nil
 }
 
@@ -369,37 +282,13 @@ func (e *compatibilityEngine) NewRangeQuery(ctx context.Context, q storage.Query
 		return nil, err
 	}
 
-	if e.debugWriter != nil {
-		explain(e.debugWriter, exec, "", "")
-	}
-
 	return &compatibilityQuery{
-		Query:       &Query{exec: exec, opts: opts},
-		engine:      e,
-		expr:        expr,
-		t:           RangeQuery,
-		debugWriter: e.debugWriter,
+		Query:  &Query{exec: exec, opts: opts},
+		engine: e,
+		expr:   expr,
+		t:      RangeQuery,
 	}, nil
 }
-
-type ExplainableQuery interface {
-	promql.Query
-
-	Explain() *ExplainOutputNode
-	Analyze() *AnalyzeOutputNode
-}
-
-type AnalyzeOutputNode struct {
-	OperatorTelemetry model.OperatorTelemetry `json:"telemetry,omitempty"`
-	Children          []AnalyzeOutputNode     `json:"children,omitempty"`
-}
-
-type ExplainOutputNode struct {
-	OperatorName string              `json:"name,omitempty"`
-	Children     []ExplainOutputNode `json:"children,omitempty"`
-}
-
-var _ ExplainableQuery = &compatibilityQuery{}
 
 type Query struct {
 	exec model.VectorOperator
@@ -419,128 +308,6 @@ func (q *Query) Analyze() *AnalyzeOutputNode {
 	return nil
 }
 
-func analyzeVector(obsv model.ObservableVectorOperator) *AnalyzeOutputNode {
-	telemetry, obsVectors := obsv.Analyze()
-
-	var children []AnalyzeOutputNode
-	for _, vector := range obsVectors {
-		children = append(children, *analyzeVector(vector))
-	}
-
-	return &AnalyzeOutputNode{
-		OperatorTelemetry: telemetry,
-		Children:          children,
-	}
-}
-
-func explainVector(v model.VectorOperator) *ExplainOutputNode {
-	name, vectors := v.Explain()
-
-	var children []ExplainOutputNode
-	for _, vector := range vectors {
-		children = append(children, *explainVector(vector))
-	}
-
-	return &ExplainOutputNode{
-		OperatorName: name,
-		Children:     children,
-	}
-}
-
-type sortOrder bool
-
-const (
-	sortOrderAsc  sortOrder = false
-	sortOrderDesc sortOrder = true
-)
-
-type resultSorter interface {
-	comparer(samples *promql.Vector) func(i, j int) bool
-}
-
-type sortFuncResultSort struct {
-	sortOrder sortOrder
-}
-
-type aggregateResultSort struct {
-	sortingLabels []string
-	groupBy       bool
-
-	sortOrder sortOrder
-}
-
-type noSortResultSort struct {
-}
-
-func newResultSort(expr parser.Expr) resultSorter {
-	switch texpr := expr.(type) {
-	case *parser.Call:
-		switch texpr.Func.Name {
-		case "sort":
-			return sortFuncResultSort{sortOrder: sortOrderAsc}
-		case "sort_desc":
-			return sortFuncResultSort{sortOrder: sortOrderDesc}
-		}
-	case *parser.AggregateExpr:
-		switch texpr.Op {
-		case parser.TOPK:
-			return aggregateResultSort{
-				sortingLabels: texpr.Grouping,
-				sortOrder:     sortOrderDesc,
-				groupBy:       !texpr.Without,
-			}
-		case parser.BOTTOMK:
-			return aggregateResultSort{
-				sortingLabels: texpr.Grouping,
-				sortOrder:     sortOrderAsc,
-				groupBy:       !texpr.Without,
-			}
-		}
-	}
-	return noSortResultSort{}
-}
-func (s noSortResultSort) comparer(samples *promql.Vector) func(i, j int) bool {
-	return func(i, j int) bool { return i < j }
-}
-
-func valueCompare(order sortOrder, l, r float64) bool {
-	if math.IsNaN(r) {
-		return true
-	}
-	if order == sortOrderAsc {
-		return l < r
-	}
-	return l > r
-}
-
-func (s sortFuncResultSort) comparer(samples *promql.Vector) func(i, j int) bool {
-	return func(i, j int) bool {
-		return valueCompare(s.sortOrder, (*samples)[i].F, (*samples)[j].F)
-	}
-}
-
-func (s aggregateResultSort) comparer(samples *promql.Vector) func(i, j int) bool {
-	return func(i int, j int) bool {
-		var iLbls labels.Labels
-		var jLbls labels.Labels
-		iLb := labels.NewBuilder((*samples)[i].Metric)
-		jLb := labels.NewBuilder((*samples)[j].Metric)
-		if s.groupBy {
-			iLbls = iLb.Keep(s.sortingLabels...).Labels()
-			jLbls = jLb.Keep(s.sortingLabels...).Labels()
-		} else {
-			iLbls = iLb.Del(s.sortingLabels...).Labels()
-			jLbls = jLb.Del(s.sortingLabels...).Labels()
-		}
-
-		lblsCmp := labels.Compare(iLbls, jLbls)
-		if lblsCmp != 0 {
-			return lblsCmp < 0
-		}
-		return valueCompare(s.sortOrder, (*samples)[i].F, (*samples)[j].F)
-	}
-}
-
 type compatibilityQuery struct {
 	*Query
 	engine     *compatibilityEngine
@@ -550,8 +317,6 @@ type compatibilityQuery struct {
 	resultSort resultSorter
 
 	cancel context.CancelFunc
-
-	debugWriter io.Writer
 }
 
 func (q *compatibilityQuery) Exec(ctx context.Context) (ret *promql.Result) {
@@ -565,9 +330,6 @@ func (q *compatibilityQuery) Exec(ctx context.Context) (ret *promql.Result) {
 	// Handle case with strings early on as this does not need us to process samples.
 	switch e := q.expr.(type) {
 	case *parser.StringLiteral:
-		if q.debugWriter != nil {
-			analyze(q.debugWriter, q.exec.(model.ObservableVectorOperator), " ", "")
-		}
 		return &promql.Result{Value: promql.String{V: e.Val, T: q.ts.UnixMilli()}}
 	}
 	ret = &promql.Result{
@@ -650,9 +412,6 @@ loop:
 			return newErrResult(ret, extlabels.ErrDuplicateLabelSet)
 		}
 		ret.Value = resultMatrix
-		if q.debugWriter != nil {
-			analyze(q.debugWriter, q.exec.(model.ObservableVectorOperator), "", "")
-		}
 		return ret
 	}
